Extracted classroom parsing logic into util file
